@@ -27,41 +27,50 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
+namespace Slim\Extras\Views;
 
 /**
- * MustacheView
+ * BlitzView
  *
- * The MustacheView is a Custom View class that renders templates using the
- * Mustache template language (http://mustache.github.com/) and the
- * [Mustache.php library](github.com/bobthecow/mustache.php).
+ * The BlitzView provides native support for the Blitz templating system
+ * for PHP. Blitz is written as C and compiled to a PHP extension. Which means
+ * it is FAST. You can learn more about Blitz at:
  *
- * There is one field that you, the developer, will need to change:
- * - mustacheDirectory
+ * <http://alexeyrybak.com/blitz/blitz_en.html>
  *
- * @package Slim
- * @author  Johnson Page <http://johnsonpage.org>
+ * The xBlitz extended blitz class provides better block handling
+ * (load assoc arrays correctly, one level)
+ *
+ * @author Tobias O. <https://github.com/tobsn>
  */
-class MustacheView extends Slim_View {
-
-    /**
-     * @var string The path to the directory containing Mustache.php
-     */
-    public static $mustacheDirectory = null;
-
-    /**
-     * Renders a template using Mustache.php.
-     *
-     * @see View::render()
-     * @param string $template The template name specified in Slim::render()
-     * @return string
-     */
-    public function render( $template ) {
-        require_once self::$mustacheDirectory . '/Mustache.php';
-        $m = new Mustache();
-        $contents = file_get_contents($this->getTemplatesDirectory() . '/' . ltrim($template, '/'));
-        return $m->render($contents, $this->data);
+class xBlitz extends \Blitz
+{
+    function xblock($k,$a)
+    {
+        foreach ($a as $v) {
+            $this->block('/' . $k, $v, true);
+        }
     }
-
 }
 
-?>
+class Blitz extends \Slim\View
+{
+    private $blitzEnvironment = null;
+
+    public function render($template)
+    {
+        $env = $this->getEnvironment($template);
+
+        return $env->parse($this->getData());
+    }
+
+    private function getEnvironment($template)
+    {
+        if (!$this->blitzEnvironment) {
+            ini_set('blitz.path', $this->getTemplatesDirectory() . '/');
+            $this->blitzEnvironment = new xBlitz($template);
+        }
+
+        return $this->blitzEnvironment;
+    }
+}
